Fixes to project files, and fixed Java errors that AIDE didn't notice.

+++ src/com/stefankopieczek/audinance/conversion/resamplers/NaiveResampler.java
@@ -22,8 +22,8 @@ public class NaiveResampler implements Resampler
 		for (int ii = 0; ii < newChannels.length; ii++)
 		{
 			newChannels[ii] = new ResamplingSource(originalChannels[ii],
-			                                       originalSampleRate,
-												   targetSampleRate);
+			                                       originalSampleRate.intValue(),
+												   targetSampleRate.intValue());
 		}
 		
 		DecodedAudio result = new DecodedAudio(newChannels,
